+++ infra/modular-examples/gcp-google-workspace/main.tf
@@ -63,7 +63,7 @@ module "google-workspace-key-secrets" {
   # source = "git::https://github.com/worklytics/psoxy//infra/modules/gcp-secrets"
 
   secret_project = var.gcp_project_id
-  secrets = {
+  secrets        = {
     "PSOXY_${replace(upper(each.key), "-", "_")}_SERVICE_ACCOUNT_KEY" : {
       value       = module.google-workspace-connection-auth[each.key].key_value
       description = "Auth key for ${each.key} service account"
@@ -110,8 +110,6 @@ module "psoxy-google-workspace-connector" {
   deployment_bundle_object_name         = module.psoxy-gcp.deployment_bundle_object_name
   path_to_config                        = "${local.base_config_path}${each.value.source_kind}.yaml"
   path_to_repo_root                     = var.psoxy_base_dir
-  salt_secret_id                        = module.psoxy-gcp.salt_secret_id
-  salt_secret_version_number            = module.psoxy-gcp.salt_secret_version_number
   example_api_calls                     = each.value.example_api_calls
   example_api_calls_user_to_impersonate = each.value.example_api_calls_user_to_impersonate
   todo_step                             = module.google-workspace-connection[each.key].next_todo_step
@@ -124,13 +122,13 @@ module "psoxy-google-workspace-connector" {
     }
   )
 
-  secret_bindings = {
+  secret_bindings = merge({
     # as SERVICE_ACCOUNT_KEY rotated by Terraform, reasonable to bind as env variable
     SERVICE_ACCOUNT_KEY = {
       secret_id      = module.google-workspace-key-secrets[each.key].secret_ids["PSOXY_${replace(upper(each.key), "-", "_")}_SERVICE_ACCOUNT_KEY"]
       version_number = module.google-workspace-key-secrets[each.key].secret_version_numbers["PSOXY_${replace(upper(each.key), "-", "_")}_SERVICE_ACCOUNT_KEY"]
     }
-  }
+  }, module.psoxy-gcp.secrets)
 }
 
 module "worklytics-psoxy-connection" {
@@ -146,35 +144,39 @@ module "worklytics-psoxy-connection" {
 }
 
 # BEGIN LONG ACCESS AUTH CONNECTORS
+locals {
+  long_access_parameters = { for entry in module.worklytics_connector_specs.enabled_oauth_secrets_to_create : "${entry.connector_name}.${entry.secret_name}" => entry }
+  long_access_parameters_by_connector = { for k, spec in module.worklytics_connector_specs.enabled_oauth_long_access_connectors :
+  k => [for secret in spec.secured_variables : "${k}.${secret.name}"]
+  }
+}
 
 resource "google_service_account" "long_auth_connector_sa" {
   for_each = module.worklytics_connector_specs.enabled_oauth_long_access_connectors
 
   project      = var.gcp_project_id
   account_id   = "psoxy-${substr(each.key, 0, 24)}"
-  display_name = "${title(each.key)}{var.connector_display_name_suffix} via Psoxy"
+  display_name = "${title(each.key)}${var.connector_display_name_suffix} via Psoxy"
 }
 
-# creates the secret, without versions.
-module "connector-long-auth-block" {
-  for_each = module.worklytics_connector_specs.enabled_oauth_long_access_connectors
+module "connector-oauth" {
+  for_each = local.long_access_parameters
 
-  source = "../../modules/gcp-oauth-long-access-strategy"
-  # source = "git::https://github.com/worklytics/psoxy//infra/modules/gcp-oauth-long-access-strategy"
+  source = "../../modules/gcp-oauth-secrets"
 
-  project_id              = var.gcp_project_id
-  function_name           = "psoxy-${each.key}"
-  token_adder_user_emails = []
+  secret_name           = "PSOXY_${upper(replace(each.value.connector_name, "-", "_"))}_${upper(each.value.secret_name)}"
+  project_id            = var.gcp_project_id
+  service_account_email = google_service_account.long_auth_connector_sa[each.value.connector_name].email
 }
 
 module "long-auth-token-secret-fill-instructions" {
-  for_each = module.worklytics_connector_specs.enabled_oauth_long_access_connectors
+  for_each = local.long_access_parameters
 
   source = "../../modules/gcp-secret-fill-md"
   # source = "git::https://github.com/worklytics/psoxy//infra/modules/gcp-secret-fill-md"
 
   project_id = var.gcp_project_id
-  secret_id  = module.connector-long-auth-block[each.key].access_token_secret_id
+  secret_id  = module.connector-oauth[each.key].secret_id
 }
 
 module "source_token_external_todo" {
@@ -187,9 +189,7 @@ module "source_token_external_todo" {
   connector_specific_external_steps = each.value.external_token_todo
   todo_step                         = 1
 
-  additional_steps = [
-    module.long-auth-token-secret-fill-instructions[each.key].todo_markdown
-  ]
+  additional_steps = [for parameter_ref in local.long_access_parameters_by_connector[each.key] : module.long-auth-token-secret-fill-instructions[parameter_ref].todo_markdown]
 }
 
 module "connector-long-auth-function" {
@@ -206,17 +206,23 @@ module "connector-long-auth-function" {
   deployment_bundle_object_name = module.psoxy-gcp.deployment_bundle_object_name
   path_to_config                = "${local.base_config_path}${each.value.source_kind}.yaml"
   path_to_repo_root             = var.psoxy_base_dir
-  salt_secret_id                = module.psoxy-gcp.salt_secret_id
-  salt_secret_version_number    = module.psoxy-gcp.salt_secret_version_number
+  example_api_calls             = each.value.example_api_calls
   todo_step                     = module.source_token_external_todo[each.key].next_todo_step
 
+<<<<<<< HEAD
   environment_variables = merge(
     var.general_environment_variables,
     try(each.value.environment_variables, {}),
+=======
+  secret_bindings = module.psoxy-gcp.secrets
+
+  environment_variables = merge(try(each.value.environment_variables, {}),
+>>>>>>> v0.4.9-rc
     {
       IS_DEVELOPMENT_MODE = contains(var.non_production_connectors, each.key)
     }
   )
+}
 
 
 
@@ -229,8 +235,6 @@ module "connector-long-auth-function" {
   #     way for users to force function restart, and env vars won't reload value of a secret until
   #     function restarts. Better to let app-code load these values from Secret Manager, cache with
   #     a TTL, and periodically refresh or refresh on auth errors.
-}
-
 module "worklytics-psoxy-connection-long-auth" {
   for_each = module.worklytics_connector_specs.enabled_oauth_long_access_connectors
 
@@ -256,13 +260,12 @@ module "psoxy-gcp-bulk" {
   worklytics_sa_emails          = var.worklytics_sa_emails
   region                        = var.gcp_region
   source_kind                   = each.value.source_kind
-  salt_secret_id                = module.psoxy-gcp.salt_secret_id
   artifacts_bucket_name         = module.psoxy-gcp.artifacts_bucket_name
   deployment_bundle_object_name = module.psoxy-gcp.deployment_bundle_object_name
-  salt_secret_version_number    = module.psoxy-gcp.salt_secret_version_number
   psoxy_base_dir                = var.psoxy_base_dir
   bucket_write_role_id          = module.psoxy-gcp.bucket_write_role_id
 
+<<<<<<< HEAD
   environment_variables = merge(
     var.general_environment_variables,
     try(each.value.environment_variables, {}),
@@ -273,3 +276,13 @@ module "psoxy-gcp-bulk" {
     }
   )
 }
+=======
+  secret_bindings = module.psoxy-gcp.secrets
+  environment_variables = {
+    SOURCE              = each.value.source_kind
+    RULES               = yamlencode(each.value.rules)
+    IS_DEVELOPMENT_MODE = contains(var.non_production_connectors, each.key)
+  }
+
+}
+>>>>>>> v0.4.9-rc