+++ infra/modules/gcp-psoxy-bulk/variables.tf
@@ -19,21 +19,6 @@ variable "source_kind" {
   description = "Kind of the content to process"
 }
 
-variable "salt_secret_id" {
-  type        = string
-  description = "Id of the secret used to salt pseudonyms"
-}
-
-variable "salt_secret_version_number" {
-  type        = string
-  description = "Version number of the secret used to salt pseudonyms"
-
-  validation {
-    condition     = can(regex("^([0-9]+)|latest$", var.salt_secret_version_number))
-    error_message = "Version number must be a number or 'latest'."
-  }
-}
-
 variable "secret_bindings" {
   type = map(object({
     secret_id      = string # NOT the full resource ID; just the secret_id within GCP project